chore(typescript): split `index.d.ts` into separate files
